feat: terms 페이지 디자인 변경, 관련 컴포넌트 추가
